+++ utils/src/utils/ConfigurationUtils.cpp
@@ -22,9 +22,9 @@
 
 namespace org::apache::nifi::minifi::utils::configuration {
 
-uint64_t getBufferSize(const Configure& configuration) {
+size_t getBufferSize(const Configure& configuration) {
   if (const auto buffer_size = configuration.get(Configure::nifi_default_internal_buffer_size); buffer_size && !buffer_size->empty()) {
-    return parsing::parseIntegral<uint64_t>(*buffer_size) | utils::orThrow(fmt::format("Invalid value '{}' for {}", *buffer_size, Configure::nifi_default_internal_buffer_size));
+    return parsing::parseIntegral<size_t>(*buffer_size) | utils::orThrow(fmt::format("Invalid value '{}' for {}", *buffer_size, Configure::nifi_default_internal_buffer_size));
   } else {
     return DEFAULT_BUFFER_SIZE;
   }
